MQGUIDE-1158 Updated the sample app for changes in 3.3.0. Moved TTS classes into sample app.
