+++ app/src/main/java/com/mapquest/navigation/sampleapp/tts/DefaultSpeechListener.java
@@ -1,0 +1,12 @@
+package com.mapquest.navigation.sampleapp.tts;
+
+public class DefaultSpeechListener implements SpeechListener {
+    @Override
+    public void onUtteranceStarted() { }
+
+    @Override
+    public void onUtteranceDone() { }
+
+    @Override
+    public void onUtteranceStopped() { }
+}
