+++ app/src/main/java/com/mapquest/navigation/sampleapp/tts/SpeechListener.java
@@ -1,0 +1,7 @@
+package com.mapquest.navigation.sampleapp.tts;
+
+public interface SpeechListener {
+    void onUtteranceStarted();
+    void onUtteranceDone();
+    void onUtteranceStopped();
+}
