+++ app/src/main/java/com/mapquest/navigation/sampleapp/tts/TextToSpeechManager.java
@@ -10,18 +10,17 @@ import android.speech.tts.UtteranceProgressListener;
 import android.support.annotation.FloatRange;
 import android.util.Log;
 
-import com.mapquest.navigation.sampleapp.tts.audio.AudioFocusLossListener;
 import com.mapquest.navigation.internal.util.ArgumentValidator;
 import com.mapquest.navigation.internal.util.LogUtil;
 import com.mapquest.navigation.internal.util.SystemVersionUtil;
-import com.mapquest.navigation.listener.SpeechListener;
+import com.mapquest.navigation.sampleapp.tts.audio.AudioFocusLossListener;
 
 import java.util.HashMap;
 import java.util.Locale;
 
+import static com.mapquest.navigation.internal.collection.CollectionsUtil.asStringMap;
 import static com.mapquest.navigation.sampleapp.tts.audio.AudioManagerUtil.getAudioManager;
 import static com.mapquest.navigation.sampleapp.tts.audio.AudioManagerUtil.wasFocusGranted;
-import static com.mapquest.navigation.internal.collection.CollectionsUtil.asStringMap;
 
 /**
  * Used to speak text. Proxies calls to Android's underlying TTS implementation, handling additional
